requirements.txt
kafka-python==2.0.2
websocket-client==1.6.0
